requirements.txt
beautifulsoup4==4.11.2
black==24.4.2
chromedriver-autoinstaller==0.6.2
flake8==7.1.0
isort==5.13.2
mypy==1.10.1
openpyxl==3.1.5
pillow==10.4.0
pyinstaller==6.9.0
pylint==3.2.5
PyQt5==5.15.9
requests==2.28.2
selenium==4.10.0
webdriver-manager==4.0.1
